Small refactor to include duck pattern
