+++ web/blog/blog.routes.js
@@ -1,0 +1,19 @@
+import validator from 'validator'
+import express from 'express'
+import {
+	getBlogPosts,
+	getBlogSeries,
+	getSingleBlogSeries,
+	getBlogPosts,
+	getPostByPermalink,
+} from './blog.controller'
+
+const router = express.Router()
+
+router.get('/', getBlogPosts)
+router.get('/series', getBlogSeries)
+router.get('/series/:permalink', getSingleBlogSeries)
+router.get('/category/:category', getBlogPosts)
+router.get('/:permalink', getPostByPermalink)
+
+module.exports = router
