+++ web/blog/contants.js
@@ -1,0 +1,1 @@
+export const MAX_PAGE_POSTS = 9
