+++ web/home/home.controller.js
@@ -3,7 +3,7 @@ const experience = require('../content/english/experience.json')
 const blogRepository = require('../repository/blog')
 
 export const getHomePage = (req, res, next) => {
-  const viewTemplateName = req.query.v1 ? 'home' : 'home_v2'
+  const viewTemplateName = req.query.v1 ? './home.view' : './home.view.v2'
   const homeContext = {
     title: 'Jorge Ferreiro - Software Engineer, Product Manager and Entrepreneur',
     image: 'https://www.ferreiro.me/images/home/welcome_to_ferreiro_v3.jpg',
